refactor: util 패키지 오타
